fs: signed timestamps, truncated to the filesystem's range as 6.8 does, never refused
The utimensat family accepts any second, as Linux does: a time before the
epoch is kept, and one past the filesystem's range is clamped to it with its
nanoseconds dropped (timestamp_truncate) — for a named node ext4's range
with 256-byte inodes (1901-12-13 to 2446-05-10), for an anonymous file
(memfd) tmpfs's every 64-bit second — where the unsigned-nanosecond ABI
refused both with EINVAL. The filesystem (fs-mem) truncates, so a WASI
guest's set times are clamped as a native guest's are. Timestamps are signed
128-bit nanoseconds across the runtime, the drivers and the trace (a JSON
number wherever 64 bits hold it; part of trace format 13), the restart
snapshot (version 7), and a (seconds, nanoseconds) pair across the C
boundary. A WASI filestat reads a time before the epoch as the epoch.

+++ crates/patina-driver-api/src/lib.rs
@@ -146,8 +146,8 @@ pub trait FsDriver: Send {
         &mut self,
         _clock: FsClock,
         _fd: Fd,
-        _atime_nanos: Option<u64>,
-        _mtime_nanos: Option<u64>,
+        _atime_nanos: Option<i128>,
+        _mtime_nanos: Option<i128>,
     ) -> DriverResult<()> {
         Err(unsupported_filesystem_operation("set times"))
     }
@@ -156,8 +156,8 @@ pub trait FsDriver: Send {
         &mut self,
         _clock: FsClock,
         _ino: u64,
-        _atime_nanos: Option<u64>,
-        _mtime_nanos: Option<u64>,
+        _atime_nanos: Option<i128>,
+        _mtime_nanos: Option<i128>,
     ) -> DriverResult<()> {
         Err(unsupported_filesystem_operation("set inode times"))
     }
@@ -165,8 +165,8 @@ pub trait FsDriver: Send {
         &mut self,
         _clock: FsClock,
         _path: &str,
-        _atime_nanos: Option<u64>,
-        _mtime_nanos: Option<u64>,
+        _atime_nanos: Option<i128>,
+        _mtime_nanos: Option<i128>,
     ) -> DriverResult<()> {
         Err(unsupported_filesystem_operation("set times by path"))
     }
